requirements.txt
absl-py==0.11.0
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alt==0.0.4
altair==6.0.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
astunparse==1.6.3
async-generator==1.10
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
binpacking==1.5.2
bitstring==4.0.1
blinker==1.9.0
bs4==0.0.1
cachetools==6.2.6
captcha==0.7.1
certifi==2026.1.4
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.8
configparser==5.3.0
contourpy==1.3.3
cryptography==46.0.4
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
DateTime==4.3
ddt==1.6.0
decorator==5.1.1
Deprecated==1.3.1
ecdsa==0.18.0
entrypoints==0.4
esptool==4.5
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.0.1
extra-streamlit-components==0.1.81
fake-useragent==1.1.1
feedparser==6.0.10
filelock==3.9.0
flatbuffers==1.12
fonttools==4.61.1
fpdf==1.7.2
frozenlist==1.8.0
future==1.0.0
gast==0.3.3
gitdb==4.0.12
GitPython==3.1.46
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==3.2.4
h11==0.16.0
h5py==3.8.0
httplib2==0.21.0
idna==3.11
importlib-metadata==5.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
looseversion==1.0.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.0
multitasking==0.0.11
narwhals==2.16.0
networkx==3.0
nibabel==5.0.1
nipype==1.8.5
numpy==2.4.2
oauthlib==3.2.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
outcome==1.2.0
packaging==26.0
pandas==2.3.3
pandas-datareader==0.10.0
parse==1.19.0
pathlib==1.0.1
pillow==12.1.0
portpicker==1.5.2
propcache==0.4.1
protobuf==6.33.5
prov==2.0.0
psutil==5.9.3
pyarrow==23.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybimstab==0.1.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydot==1.4.2
pyee==8.2.2
Pygments==2.13.0
PyJWT==2.11.0
Pympler==1.0.1
PyMuPDF==1.26.7
pyodbc==5.3.0
pyparsing==3.3.2
pypdf==3.4.1
PyPDF2==3.0.1
pypng==0.20220715.0
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.5
PyYAML==6.0.3
pyzbar==0.1.9
qrcode==8.2
rdflib==6.2.0
reedsolo==1.6.0
referencing==0.37.0
reportlab==4.4.9
requests==2.32.5
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==12.6.0
rpds-py==0.30.0
rsa==4.9
scipy==1.17.0
selenium==4.16.0
semver==2.13.0
sgmllib3k==1.0.0
Shapely==1.8.5.post1
simplejson==3.18.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==1.4.43
starlette==0.50.0
stauth==0.1.7
streamlit==1.54.0
streamlit-authenticator==0.2.2
svglib==1.4.1
tdqm==0.0.1
tenacity==9.1.4
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.5.4
tqdm==4.67.3
traits==6.3.2
trio==0.22.0
trio-websocket==0.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==4.2
urllib3==2.6.3
uvicorn==0.22.0
validators==0.20.0
w3lib==2.1.1
watchdog==6.0.0
webencodings==0.5.1
websockets==10.4
wrapt==2.0.1
wsproto==1.2.0
xlrd==2.0.1
yarl==1.22.0
zbarlight==4.0
zipp==3.8.1
zope.interface==5.5.0
